requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
requests>=2.31.0
anthropic>=0.25.0
openai>=1.14.0
pandas>=2.2.0
python-dotenv>=1.0.0
